Separate the example
