+++ src/soundness/compositionality/mod.rs
@@ -1,4 +1,5 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
+pub mod examples;
 pub mod proof;
 pub mod state_machine;
